refactor: move formatScore to Calculator, import into CommonSpace
