+++ app/src/Components/Calculator.tsx
@@ -1,7 +1,22 @@
 import { useState } from 'react';
 import type { CSSProperties } from 'react';
 import { Icon } from './Icons';
-import { formatScore } from './CommonSpace';
+
+// Format a score for display — raw up to 99,999, then SI prefixes (truncated), then scientific
+export const formatScore = (n: number): string => {
+  const abs = Math.abs(n);
+  const sign = n < 0 ? '-' : '';
+  if (abs < 1e5) return String(n);
+  const fmt = (val: number, suffix: string) => {
+    const floored = Math.floor(val * 100) / 100;
+    return `${sign}${floored.toFixed(2).replace(/\.?0+$/, '')}${suffix}`;
+  };
+  if (abs < 1e6) return `${sign}${Math.floor(abs / 1e3)}k`;
+  if (abs < 1e9) return fmt(abs / 1e6, 'M');
+  if (abs < 1e12) return fmt(abs / 1e9, 'G');
+  if (abs < 1e15) return fmt(abs / 1e12, 'T');
+  return `${sign}${abs.toExponential(3).replace('e+', ' e').replace('e-', ' e-').replace('e', ' e')}`;
+};
 
 interface CalculatorProps {
   initialValue: number;
